Rubocop fix

+++ app/jobs/cart_cleaner.rb
@@ -3,8 +3,6 @@ class CartCleaner < ApplicationJob
 
   def perform(order_id)
     order = Order.find(order_id)
-    if order.status.eql?('initialized') || (order.last_action - DateTime.now).eql?(168)
-      order.destroy
-    end
+    order.destroy if order.status.eql?('initialized') || (order.last_action - DateTime.now).eql?(168)
   end
 end
